refactor: BaseBallGame 클래스의 이름 Computer로 변경
랜덤 숫자를 생성하고 비교하는 과정은 컴퓨터가 수행하고, BaseballGame은 게임 진행에 어울리는 이름이라고 생각했다. 그래서 BaseballGame 클래스를 다른 객체를 위해 사용하기 위해 랜덤 번호 생성은 Computer로 바꿔주었다.
